+++ src/nl_sgtk/nl_sgtk.py
@@ -17,7 +17,7 @@ from urllib.parse import parse_qs, urlparse
 log = logging.getLogger(__name__)
 
 # Keep a module version to align with setup.py
-__version__ = "0.9.2"
+__version__ = "0.11.0"
 
 try:
     notify_if_update_available(__version__)